Add a Jekyll format option to the show command
Jekyll is a static site generator based on YAML and markdown. This
output can be used to generate crate pages for a simple Alire website.

The main change of this patch is to introduce a "Yamlable" interface to
describe objects that can produce a text YAML representation, and then
implement this interface for the required types.